hydroviewer hiwat
Hidroviewer Hiwat is a web application that was created to allow users to visualize results from the Hiwat (High-Impact Weather Assessment Toolkit) model. In addition, the application displays only results from rivers that are not transboundary. The app also provides historical data, and the ability to download the forecast and historical data.